Sanitize Playernames to prevent Database insertion errors
Filter out "SQL compatible" Characters from Playername to prevent failures when inserting Identifiers to the Database.
Use a default values on Names that are completely "custom" to have a value.

This does not affect the display of online players which in itself is fully working with custom characters, it only fixes offline banning users with custom names that previously would not have been inserted into the Database.